+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.60.0] - 2021-02-02
+
+# Added
+
+- Added class weights support for Longformer classification
+- Added new classification models:
+  - SqueezeBert
+  - DeBERTa
+  - MPNet
+
+# Changed
+
+- Updated ClassificationModel logic to make it easier to add new models
+
 ## [0.51.16] - 2021-01-29
 
 ## Fixed
@@ -1386,7 +1400,11 @@ Model checkpoint is now saved for all epochs again.
 
 - This CHANGELOG file to hopefully serve as an evolving example of a standardized open source project CHANGELOG.
 
-[0.51.15]: https://github.com/ThilinaRajapakse/simpletransformers/compare/2af55e9...HEAD
+[0.60.0]: https://github.com/ThilinaRajapakse/simpletransformers/compare/5840749...HEAD
+
+[0.51.16]: https://github.com/ThilinaRajapakse/simpletransformers/compare/b42898e...5840749
+
+[0.51.15]: https://github.com/ThilinaRajapakse/simpletransformers/compare/2af55e9...b42898e
 
 [0.51.14]: https://github.com/ThilinaRajapakse/simpletransformers/compare/278fca1...2af55e9